Fix numpy boolean subtract error in Factor Analysis
- Convert pd.get_dummies output to int in regime_factors.py
- Add boolean-to-int conversion in scenario_detector.detect()
- Add boolean-to-int conversion in normalizers._zscore_normalize()
- Add boolean-to-int conversion in outlier_handler for z-score calculation

These changes prevent the "numpy boolean subtract, the `-` operator,
is not supported" error that occurs with one-hot encoded columns
from pd.get_dummies() in newer numpy versions.

+++ Classes/FactorAnalysis/factors/regime_factors.py
@@ -284,10 +284,10 @@ class RegimeFactors:
         """
         df = factors_df.copy()
 
-        # One-hot encode regimes
+        # One-hot encode regimes (convert to int to avoid numpy boolean subtract error)
         for col in ['regime_volatility', 'regime_trend', 'regime_momentum']:
             if col in df.columns:
-                dummies = pd.get_dummies(df[col], prefix=col)
+                dummies = pd.get_dummies(df[col], prefix=col).astype(int)
                 df = pd.concat([df, dummies], axis=1)
 
         return df
